test for expandAbbreviations. move InputExpansion
