+++ rabbit-escape-ui-text/src/rabbitescape/ui/text/InputExpansion.java
@@ -1,4 +1,4 @@
-package rabbitescape.engine.textworld;
+package rabbitescape.ui.text;
 
 public class InputExpansion
 {
